Add an option to keep temp files

+++ backup.sh
@@ -158,7 +158,7 @@ compare()
 	wait
 
 	# clean up
-	rm "$BACKUP_TMP".sql "$BACKUP_TMP".files
+  test -z "$BACKUP_KEEP_TMP" && rm "$BACKUP_TMP".sql "$BACKUP_TMP".files
 }
 
 ### RUN ###
